Be careful with null pointers from DirectSound
IDirectSoundBuffer::Lock may return null pointers for either locked
regions. The second case is documented so the lack of p1 check was a
mistake. However, the first is specifically documented *not* to return a
null pointer. This is probably a bug in DirectSound, but it's easy to
work around.

(IMHO, it's also a serious defect in the C standard that memset does not
accept a null pointer with a zero length, which is the only reason this
matters in the first place.)